requirements.txt
pdfminer.six
reportlab
pytest